Extract self-extracting archives for searchability
Where possible, archive contents were left in the same directory where
the archive itself existed.

Several archives contained files with ambiguous names such as "README"
and editorially it made more sense to make a subdirectory to extract
those into.  This is probably the most controversial part of the commit
as it is a small structural change.

EMSIG.400 in the 03-TELCOM was moved.  The self-extracting archive
contained a byte identical file and I elected instead to keep that copy
in EMSIG/ instead of the existing file in the flat directory.