Unpack ARC files in cases where the ARC contents are useful
Most of the ARC files in the archive are just compressed bundles of
stuff that's already unpacked.  Some, though, are distribution packages
of things that have interesting docs or discussion that isn't
searchable.

In cases where the contents of an ARC aren't already present and
accessible, those things are unpacked.  In cases where unpacking the ARC
just creates duplicate or near-duplicate files, nothing was done.